Currency

+++ app/Http/Controllers/MasterSetupController.php
@@ -5,7 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Designation;
 use App\Department;
-use App\project;
+use App\Currency;
+use App\Project;
 use App\Branch;
 use Auth;
 
@@ -45,8 +46,12 @@ class MasterSetupController extends Controller
 
     public function department_delete($id) {
         $department = Department::find($id);
-        $department->delete();
-        return redirect('departments')->with('message','Department Deleted Successfully!');
+        if($department->company_id == Auth::user()->company_id){
+            $department->delete();
+            return redirect('departments')->with('message','Department Deleted Successfully!');
+        }else{
+            return redirect('departments')->with('message','Do not try to be too smart!');
+        }
     }
 
     public function designation_index() {
@@ -78,8 +83,12 @@ class MasterSetupController extends Controller
 
     public function designation_delete($id) {
         $designation = Designation::find($id);
-        $designation->delete();
-        return redirect('designations')->with('message','Designation Deleted Successfully!');
+        if($designation->company_id == Auth::user()->company_id){
+            $designation->delete();
+            return redirect('designations')->with('message','Designation Deleted Successfully!');
+        }else{
+            return redirect('designations')->with('message','Do not try to be too smart!');
+        }
     }
 
 
@@ -114,8 +123,12 @@ class MasterSetupController extends Controller
 
     public function project_delete($id) {
         $project = Project::find($id);
-        $project->delete();
-        return redirect('projects')->with('message','Project Deleted Successfully!');
+        if($project->company_id == Auth::user()->company_id){
+            $project->delete();
+            return redirect('projects')->with('message','Project Deleted Successfully!');
+        }else{
+            return redirect('projects')->with('message','Do not try to be too smart!');
+        }
     }
 
 
@@ -157,4 +170,74 @@ class MasterSetupController extends Controller
             return redirect('branches')->with('message','Do not try to be too smart!');
         }
     }
+
+
+    public function currency_index() {
+        $currencies = Currency::orderBy('id','asc')->paginate(10);
+        return view('master_setup.currencies',compact('currencies'));
+    }
+
+    public function currency_add(Request $request) {
+        $currency = new Currency;
+        $currency->currency_name    = $request->currency_name;
+        $currency->full_unit_name   = $request->full_unit_name;
+        $currency->sub_unit_name    = $request->sub_unit_name;
+        $currency->save();
+        return redirect('currencies')->with('message','Currency Added Successfully!');
+    }
+
+    public function currency_get($id) {
+        $currency = Currency::where('id',$id)->first();
+        echo $currency;
+    }
+
+    public function currency_update(Request $request,$id) {
+        $currency = Currency::where('id',$id)->first();
+        $currency->currency_name    = $request->currency_name;
+        $currency->full_unit_name   = $request->full_unit_name;
+        $currency->sub_unit_name    = $request->sub_unit_name;
+        $currency->save();
+        return redirect('currencies')->with('message','Currency Updated Successfully!');
+    }
+
+    public function currency_delete($id) {
+        $currency = Currency::find($id);
+        $currency->delete();
+        return redirect('currencies')->with('message','Currency Deleted Successfully!');
+    }
+
+    /*public function bank_account_index() {
+        $projects = Project::where('company_id',Auth::user()->company_id)->orderBy('name','asc')->paginate(10);
+        return view('master_setup.projects',compact('projects'));
+    }
+
+    public function bank_account_add(Request $request) {
+        $project = new Project;
+        $project->company_id    = Auth::user()->company_id;
+        $project->name          = $request->name;
+        $project->project_id    = $request->project_id;
+        $project->address       = $request->address;
+        $project->save();
+        return redirect('projects')->with('message','Project Added Successfully!');
+    }
+
+    public function bank_account_get($id) {
+        $project = Project::where('id',$id)->first();
+        echo $project;
+    }
+
+    public function bank_account_update(Request $request,$id) {
+        $project = Project::where('id',$id)->first();
+        $project->name          = $request->name;
+        $project->project_id    = $request->project_id;
+        $project->address       = $request->address;
+        $project->save();
+        return redirect('projects')->with('message','Project Updated Successfully!');
+    }
+
+    public function bank_account_delete($id) {
+        $project = Project::find($id);
+        $project->delete();
+        return redirect('projects')->with('message','Project Deleted Successfully!');
+    }*/
 }
